Added return call to fix appcenter compile error

+++ Pulse/Sources/PulseUI/Features/Console/ConsoleNetworkRequestView.swift
@@ -97,7 +97,7 @@ struct ConsoleNetworkRequestView: View {
         #if os(tvOS)
         20
         #else
-        return 10
+        return (CGFloat)10
         #endif
     }
 }
